+++ neurons/validators/score_manager.py
@@ -4,11 +4,6 @@ import numpy as np
 
 from typing import List
 
-
-from webgenie.base.utils.weight_utils import (
-    process_weights_for_netuid,
-    convert_weights_and_uids_for_emit,
-) 
 from webgenie.base.neuron import BaseNeuron
 from webgenie.constants import CONSIDERING_SESSION_COUNTS
 from webgenie.storage import send_challenge_to_stats_collector
@@ -109,15 +104,9 @@ class ScoreManager:
         self.winners[-1] = np.argmax(self.session_accumulated_scores)
         bt.logging.info(f"Updated winners: {self.winners}")
         self.should_save = True
-    
-    def set_weights(self):
-        """
-        Sets the validator weights to the metagraph hotkeys based on the scores it has received from the miners. The weights determine the trust and incentive level the validator assigns to miner nodes on the network.
-        """
-        
+
+    def send_challenge_to_stats_collector(self):
         with self.lock:
-            if not self.neuron.should_set_weights():
-                return
             current_session = self.neuron.session
 
         if current_session != self.last_send_stats_collector_session:
@@ -127,56 +116,10 @@ class ScoreManager:
             except Exception as e:
                 bt.logging.error(f"Error sending challenge to stats collector: {e}")
         
+    def get_scores(self):
         with self.lock:
             scores = np.zeros(self.neuron.metagraph.n, dtype=np.float32)
             for winner in self.winners:
                 scores[winner] += 1
 
-
-        # Calculate the average reward for each uid across non-zero values.
-        # Replace any NaN values with 0.
-        # Compute the norm of the scores
-        norm = np.linalg.norm(scores, ord=1, axis=0, keepdims=True)
-
-        # Check if the norm is zero or contains NaN values
-        if np.any(norm == 0) or np.isnan(norm).any():
-            norm = np.ones_like(norm)  # Avoid division by zero or NaN
-
-        # Compute raw_weights safely
-        raw_weights = scores / norm
-        
-        with self.lock:
-            # Process the raw weights to final_weights via subtensor limitations.
-            (
-                processed_weight_uids,
-                processed_weights,
-            ) = process_weights_for_netuid(
-                uids=self.neuron.metagraph.uids,
-                weights=raw_weights,
-                netuid=self.neuron.config.netuid,
-                subtensor=self.neuron.subtensor,
-                metagraph=self.neuron.metagraph,
-            )
-
-            # Convert to uint16 weights and uids.
-            (
-                uint_uids,
-                uint_weights,
-            ) = convert_weights_and_uids_for_emit(
-                uids=processed_weight_uids, weights=processed_weights
-            )
-            # Set the weights on chain via our subtensor connection.
-            result, msg = self.neuron.subtensor.set_weights(
-                wallet=self.neuron.wallet,
-                netuid=self.neuron.config.netuid,
-                uids=uint_uids,
-                weights=uint_weights,
-                wait_for_finalization=False,
-                wait_for_inclusion=False,
-                version_key=self.neuron.spec_version,
-            )
-            if result is True:
-                bt.logging.success("set_weights on chain successfully!")
-            else:
-                bt.logging.error("set_weights failed", msg)
-                
+        return scores
